pyproject.toml
[project]
name = "hndigest"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiohttp",
    "fastapi>=0.110.0",
    "pydantic>=2.0",
    "PyYAML",
    "trafilatura",
    "python-dotenv",
    "uvicorn[standard]",
]

[project.optional-dependencies]
dev = [
    "httpx",
    "pytest",
    "pytest-asyncio",
    "python-dotenv",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
